replication: transform a possible issue to a fatal issue
In a weird case, the simple_query in the replication_slot_create would
return a E42601 (syntax error). I'm not sure why, but we'll now crash
if that's the case instead of continuing and having the WSS server
being useless (but up).

Also "fix" a possible same issue as the above one (unwrap crashing but
the server continue doing is life as if everything was ok).
CopyBothSimple will now crash if it cannot get the Duplex stream.

also update to config 0.12